@@ -1,7 +1,7 @@
 [package]
-name = "backend"
+name = "catcolab-backend"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
 [dependencies]
 axum = "0.7.7"
