@@ -2,7 +2,7 @@
 name = "catlog-wasm"
 version = "0.1.0"
 authors = ["Evan Patterson <evan@epatters.org>"]
-edition = "2021"
+edition = "2024"
 
 [lib]
 crate-type = ["cdylib", "rlib"]
