@@ -4,8 +4,6 @@ import { NodeWSServerAdapter } from "@automerge/automerge-repo-network-websocket
 import express from "express";
 import * as ws from "ws";
 
-import type { JsonValue } from "../../backend/pkg/src/index.ts";
-
 export class AutomergeServer {
     private docMap: Map<string, DocHandle<unknown>>;
     private app: express.Express;
@@ -13,7 +11,7 @@ export class AutomergeServer {
     private wss: ws.WebSocketServer;
     private repo: Repo;
 
-    public handleChange?: (refId: string, content: JsonValue) => void;
+    public handleChange?: (refId: string, content: any) => void;
 
     constructor(port: number | string) {
         this.docMap = new Map();
