@@ -1,8 +1,9 @@
 //! Double category theory and two-dimensional categorical logic.
 
-pub mod computad;
-pub mod diagram;
-pub mod pasting;
+pub mod category;
+pub mod graph;
+pub mod tree;
+pub mod tree_algorithms;
 
 pub mod model;
 pub mod model_diagram;
