@@ -15,11 +15,11 @@ use crate::validate::{self, Validate};
 
 /** A functional mapping.
 
-A mapping takes values of type [`Dom`](Self::Dom) to values of type
+A mapping sends values of type [`Dom`](Self::Dom) to values of type
 [`Cod`](Self::Cod). Unlike a function, a mapping need not be defined on its
 whole domain. A mapping is thus more like a partial function, but it does not
-actually know its domain of definition. If needed, that information should be
-provided separately, preferably as a [`Set`].
+actually know its intended domain of definition. If needed, that information
+should be provided separately, preferably as a [`Set`].
 
 Neither the domain nor the codomain of the mapping are assumed to be finite.
  */
@@ -31,7 +31,31 @@ pub trait Mapping {
     type Cod: Eq + Clone;
 
     /// Applies the mapping at a point possibly in the domain.
-    fn apply(&self, x: &Self::Dom) -> Option<&Self::Cod>;
+    fn apply(&self, x: &Self::Dom) -> Option<Self::Cod>;
+
+    /** Is the mapping defined at a point?
+
+    The default implementation checks whether [`apply`](Self::apply) returns
+    something. Often this method can be given a more efficient implementation
+    that avoids allocating.
+    */
+    fn is_set(&self, x: &Self::Dom) -> bool {
+        self.apply(x).is_some()
+    }
+}
+
+/** A mutable [mapping](Mapping).
+
+Besides being mutable, such a mapping is also assumed to own its values (how
+else could they be mutated?) and thus also allows access by reference.
+ */
+pub trait MutMapping: Mapping {
+    /** Gets the value of the mapping at a point possibly in the domain.
+
+    The same as [`apply`](Mapping::apply) but returns by reference rather than
+    by value.
+    */
+    fn get(&self, x: &Self::Dom) -> Option<&Self::Cod>;
 
     /** Sets the mapping at a point.
 
@@ -54,11 +78,6 @@ pub trait Mapping {
             Some(y) => self.set(x, y),
             None => self.unset(&x),
         }
-    }
-
-    /// Is the mapping defined at a point?
-    fn is_set(&self, x: &Self::Dom) -> bool {
-        self.apply(x).is_some()
     }
 }
 
@@ -106,11 +125,13 @@ where
     Cod: Set<Elem = Map::Cod>,
 {
     /// Iterates over failures to be a function.
-    pub fn iter_invalid(&self) -> impl Iterator<Item = InvalidFunction<Map::Dom>> + 'a {
+    pub fn iter_invalid(
+        &self,
+    ) -> impl Iterator<Item = InvalidFunction<Map::Dom>> + 'a + use<'a, Map, Dom, Cod> {
         let Function(mapping, dom, cod) = self;
         dom.iter().filter_map(|x| match mapping.apply(&x) {
             Some(y) => {
-                if cod.contains(y) {
+                if cod.contains(&y) {
                     None
                 } else {
                     Some(InvalidFunction::Cod(x))
@@ -173,7 +194,17 @@ impl<T: Eq + Clone> Mapping for VecColumn<T> {
     type Dom = usize;
     type Cod = T;
 
-    fn apply(&self, i: &usize) -> Option<&T> {
+    fn apply(&self, i: &usize) -> Option<T> {
+        self.get(i).cloned()
+    }
+
+    fn is_set(&self, i: &usize) -> bool {
+        *i < self.0.len() && self.0[*i].is_some()
+    }
+}
+
+impl<T: Eq + Clone> MutMapping for VecColumn<T> {
+    fn get(&self, i: &usize) -> Option<&T> {
         if *i < self.0.len() {
             self.0[*i].as_ref()
         } else {
@@ -194,10 +225,6 @@ impl<T: Eq + Clone> Mapping for VecColumn<T> {
         } else {
             None
         }
-    }
-
-    fn is_set(&self, i: &usize) -> bool {
-        *i < self.0.len() && self.0[*i].is_some()
     }
 }
 
@@ -236,7 +263,21 @@ where
     type Dom = K;
     type Cod = V;
 
-    fn apply(&self, x: &K) -> Option<&V> {
+    fn apply(&self, x: &K) -> Option<V> {
+        self.0.get(x).cloned()
+    }
+    fn is_set(&self, x: &K) -> bool {
+        self.0.contains_key(x)
+    }
+}
+
+impl<K, V, S> MutMapping for HashColumn<K, V, S>
+where
+    K: Eq + Hash + Clone,
+    V: Eq + Clone,
+    S: BuildHasher,
+{
+    fn get(&self, x: &K) -> Option<&V> {
         self.0.get(x)
     }
     fn set(&mut self, x: K, y: V) -> Option<V> {
@@ -244,9 +285,6 @@ where
     }
     fn unset(&mut self, x: &K) -> Option<V> {
         self.0.remove(x)
-    }
-    fn is_set(&self, x: &K) -> bool {
-        self.0.contains_key(x)
     }
 }
 
@@ -400,18 +438,28 @@ impl<Dom, Cod, Col, Ind> Mapping for IndexedColumn<Dom, Cod, Col, Ind>
 where
     Dom: Eq + Clone,
     Cod: Eq + Clone,
-    Col: Column<Dom = Dom, Cod = Cod>,
-    Ind: Index<Dom = Dom, Cod = Cod>,
+    Col: Mapping<Dom = Dom, Cod = Cod>,
 {
     type Dom = Dom;
     type Cod = Cod;
 
-    fn apply(&self, x: &Dom) -> Option<&Cod> {
+    fn apply(&self, x: &Dom) -> Option<Cod> {
         self.mapping.apply(x)
     }
-
     fn is_set(&self, x: &Dom) -> bool {
         self.mapping.is_set(x)
+    }
+}
+
+impl<Dom, Cod, Col, Ind> MutMapping for IndexedColumn<Dom, Cod, Col, Ind>
+where
+    Dom: Eq + Clone,
+    Cod: Eq + Clone,
+    Col: MutMapping<Dom = Dom, Cod = Cod>,
+    Ind: Index<Dom = Dom, Cod = Cod>,
+{
+    fn get(&self, x: &Dom) -> Option<&Cod> {
+        self.mapping.get(x)
     }
 
     fn set(&mut self, x: Dom, y: Cod) -> Option<Cod> {
@@ -474,17 +522,23 @@ impl SkelIndexedColumn {
 impl Mapping for SkelIndexedColumn {
     type Dom = usize;
     type Cod = usize;
-    fn apply(&self, x: &usize) -> Option<&usize> {
+    fn apply(&self, x: &usize) -> Option<usize> {
         self.0.apply(x)
+    }
+    fn is_set(&self, x: &usize) -> bool {
+        self.0.is_set(x)
+    }
+}
+
+impl MutMapping for SkelIndexedColumn {
+    fn get(&self, x: &usize) -> Option<&usize> {
+        self.0.get(x)
     }
     fn set(&mut self, x: usize, y: usize) -> Option<usize> {
         self.0.set(x, y)
     }
     fn unset(&mut self, x: &usize) -> Option<usize> {
         self.0.unset(x)
-    }
-    fn is_set(&self, x: &usize) -> bool {
-        self.0.is_set(x)
     }
 }
 
@@ -526,17 +580,23 @@ impl<T: Eq + Hash + Clone> IndexedVecColumn<T> {
 impl<T: Eq + Hash + Clone> Mapping for IndexedVecColumn<T> {
     type Dom = usize;
     type Cod = T;
-    fn apply(&self, x: &usize) -> Option<&T> {
+    fn apply(&self, x: &usize) -> Option<T> {
         self.0.apply(x)
+    }
+    fn is_set(&self, x: &usize) -> bool {
+        self.0.is_set(x)
+    }
+}
+
+impl<T: Eq + Hash + Clone> MutMapping for IndexedVecColumn<T> {
+    fn get(&self, x: &usize) -> Option<&T> {
+        self.0.get(x)
     }
     fn set(&mut self, x: usize, y: T) -> Option<T> {
         self.0.set(x, y)
     }
     fn unset(&mut self, x: &usize) -> Option<T> {
         self.0.unset(x)
-    }
-    fn is_set(&self, x: &usize) -> bool {
-        self.0.is_set(x)
     }
 }
 
@@ -577,17 +637,28 @@ where
 {
     type Dom = K;
     type Cod = V;
-    fn apply(&self, x: &K) -> Option<&V> {
+    fn apply(&self, x: &K) -> Option<V> {
         self.0.apply(x)
+    }
+    fn is_set(&self, x: &K) -> bool {
+        self.0.is_set(x)
+    }
+}
+
+impl<K, V, S> MutMapping for IndexedHashColumn<K, V, S>
+where
+    K: Eq + Hash + Clone,
+    V: Eq + Hash + Clone,
+    S: BuildHasher,
+{
+    fn get(&self, x: &K) -> Option<&V> {
+        self.0.get(x)
     }
     fn set(&mut self, x: K, y: V) -> Option<V> {
         self.0.set(x, y)
     }
     fn unset(&mut self, x: &K) -> Option<V> {
         self.0.unset(x)
-    }
-    fn is_set(&self, x: &K) -> bool {
-        self.0.is_set(x)
     }
 }
 
@@ -621,7 +692,8 @@ mod tests {
         let mut col = VecColumn::new(vec!["foo", "bar", "baz"]);
         assert!(!col.is_empty());
         assert!(col.is_set(&2));
-        assert_eq!(col.apply(&2), Some(&"baz"));
+        assert_eq!(col.get(&2), Some(&"baz"));
+        assert_eq!(col.apply(&2), Some("baz"));
         assert_eq!(col.apply(&3), None);
         assert_eq!(col.update(2, None), Some("baz"));
         assert!(!col.is_set(&2));
@@ -640,7 +712,8 @@ mod tests {
         col.set('b', "bar");
         col.set('c', "baz");
         assert!(!col.is_empty());
-        assert_eq!(col.apply(&'c'), Some(&"baz"));
+        assert_eq!(col.get(&'c'), Some(&"baz"));
+        assert_eq!(col.apply(&'c'), Some("baz"));
         assert_eq!(col.unset(&'c'), Some("baz"));
         assert!(!col.is_set(&'c'));
         col.set('c', "bar");
@@ -655,7 +728,8 @@ mod tests {
         let mut col = SkelIndexedColumn::new(&[1, 3, 5]);
         assert!(!col.is_empty());
         assert!(col.is_set(&2));
-        assert_eq!(col.apply(&2), Some(&5));
+        assert_eq!(col.get(&2), Some(&5));
+        assert_eq!(col.apply(&2), Some(5));
         let preimage: Vec<_> = col.preimage(&5).collect();
         assert_eq!(preimage, vec![2]);
 
@@ -671,7 +745,7 @@ mod tests {
         let mut col = IndexedVecColumn::new(&["foo", "bar", "baz"]);
         assert!(!col.is_empty());
         assert!(col.is_set(&2));
-        assert_eq!(col.apply(&2), Some(&"baz"));
+        assert_eq!(col.apply(&2), Some("baz"));
         let preimage: Vec<_> = col.preimage(&"baz").collect();
         assert_eq!(preimage, vec![2]);
 
@@ -690,7 +764,7 @@ mod tests {
         col.set('b', "bar");
         col.set('c', "baz");
         assert!(!col.is_empty());
-        assert_eq!(col.apply(&'c'), Some(&"baz"));
+        assert_eq!(col.apply(&'c'), Some("baz"));
         let preimage: Vec<_> = col.preimage(&"baz").collect();
         assert_eq!(preimage, vec!['c']);
 
