@@ -10,13 +10,13 @@ export function FormGroup(props: {
     return <dl class={props.compact ? "compact-form-group" : "form-group"}>{props.children}</dl>;
 }
 
-/** Text input field in a form group. */
-export function TextInputField(
-    allProps: {
-        label: string | JSX.Element;
-        error?: string;
-    } & Omit<ComponentProps<"input">, "id">,
-) {
+type InputFieldProps = {
+    label: string | JSX.Element;
+    error?: string;
+};
+
+/** Input field in a form group. */
+export function InputField(allProps: InputFieldProps & Omit<ComponentProps<"input">, "id">) {
     const fieldId = createUniqueId();
 
     const [props, inputProps] = splitProps(allProps, ["label", "error"]);
@@ -27,18 +27,19 @@ export function TextInputField(
                 <label for={fieldId}>{props.label}</label>
             </dt>
             <dd>
-                <input {...inputProps} id={fieldId} type="text" />
-                <InputError error={props.error} />
+                <input {...inputProps} id={fieldId} />
+                <FieldError error={props.error} />
             </dd>
         </>
     );
 }
 
-const InputError = (props: { error?: string }) => (
-    <Show when={props.error}>
-        <div class="error">{props.error}</div>
-    </Show>
-);
+/** Text input field in a form group. */
+export function TextInputField(
+    props: InputFieldProps & Omit<ComponentProps<"input">, "id" | "type">,
+) {
+    return <InputField type="text" {...props} />;
+}
 
 /** Select field in a form group.
 
@@ -75,3 +76,29 @@ export function SelectField(
         </>
     );
 }
+
+/** Text area field in a form group. */
+export function TextAreaField(allProps: InputFieldProps & Omit<ComponentProps<"textarea">, "id">) {
+    const fieldId = createUniqueId();
+
+    const [props, textProps] = splitProps(allProps, ["label", "error"]);
+
+    return (
+        <>
+            <dt>
+                <label for={fieldId}>{props.label}</label>
+            </dt>
+            <dd>
+                <textarea id={fieldId} {...textProps} />
+                <FieldError error={props.error} />
+            </dd>
+        </>
+    );
+}
+
+/** Validation error for a field. */
+const FieldError = (props: { error?: string }) => (
+    <Show when={props.error}>
+        <div class="error">{props.error}</div>
+    </Show>
+);
