@@ -2,14 +2,14 @@ import { DropdownMenu } from "@kobalte/core/dropdown-menu";
 import { useNavigate } from "@solidjs/router";
 import { getAuth, signOut } from "firebase/auth";
 import { useAuth, useFirebaseApp } from "solid-firebase";
-import { type JSX, Show, createSignal, useContext } from "solid-js";
+import { type JSX, Show, useContext } from "solid-js";
 import invariant from "tiny-invariant";
 
 import { useApi } from "../api";
-import { Dialog, IconButton } from "../components";
+import { IconButton } from "../components";
 import { createModel } from "../model/document";
 import { TheoryLibraryContext } from "../stdlib";
-import { Login } from "../user";
+import { PageActionsContext } from "./context";
 
 import FilePlus from "lucide-solid/icons/file-plus";
 import Info from "lucide-solid/icons/info";
@@ -17,6 +17,7 @@ import LogInIcon from "lucide-solid/icons/log-in";
 import LogOutIcon from "lucide-solid/icons/log-out";
 import MenuIcon from "lucide-solid/icons/menu";
 import SettingsIcon from "lucide-solid/icons/settings";
+import UploadIcon from "lucide-solid/icons/upload";
 
 import "./menubar.css";
 
@@ -56,28 +57,22 @@ export function AppMenu(props: {
     const firebaseApp = useFirebaseApp();
     const auth = useAuth(getAuth(firebaseApp));
 
-    const [loginOpen, setLoginOpen] = createSignal(false);
-
     // Root the dialog here so that it is not destroyed when the menu closes.
     return (
         <>
-            <HamburgerMenu>
+            <HamburgerMenu disabled={props.disabled}>
                 {props.children}
                 <Show when={props.children}>
                     <MenuSeparator />
                 </Show>
+
                 <HelpMenuItem />
-                <Show
-                    when={auth.data}
-                    fallback={<LogInMenuItem showLogin={() => setLoginOpen(true)} />}
-                >
+
+                <Show when={auth.data} fallback={<LogInMenuItem />}>
                     <SettingsMenuItem />
                     <LogOutMenuItem />
                 </Show>
             </HamburgerMenu>
-            <Dialog open={loginOpen()} onOpenChange={setLoginOpen} title="Log in">
-                <Login onComplete={() => setLoginOpen(false)} />
-            </Dialog>
         </>
     );
 }
@@ -86,6 +81,7 @@ export function AppMenu(props: {
 export const DefaultAppMenu = () => (
     <AppMenu>
         <NewModelItem />
+        <ImportMenuItem />
     </AppMenu>
 );
 
@@ -110,6 +106,19 @@ export function NewModelItem() {
     );
 }
 
+/** Menu item to import a document. */
+export function ImportMenuItem() {
+    const actions = useContext(PageActionsContext);
+    invariant(actions, "Page actions should be provided");
+
+    return (
+        <MenuItem onSelect={actions.showImportDialog}>
+            <UploadIcon />
+            <MenuItemLabel>{"Import notebook"}</MenuItemLabel>
+        </MenuItem>
+    );
+}
+
 /** Menu item navigating to the top-level application help. */
 function HelpMenuItem() {
     const navigate = useNavigate();
@@ -122,11 +131,12 @@ function HelpMenuItem() {
     );
 }
 
-function LogInMenuItem(props: {
-    showLogin: () => void;
-}) {
+function LogInMenuItem() {
+    const actions = useContext(PageActionsContext);
+    invariant(actions, "Page actions should be provided");
+
     return (
-        <MenuItem onSelect={props.showLogin}>
+        <MenuItem onSelect={actions.showLoginDialog}>
             <LogInIcon />
             <MenuItemLabel>{"Log in or sign up"}</MenuItemLabel>
         </MenuItem>
